+++ epsilon/apps/shared/interactive_curve_view_controller.h
@@ -101,7 +101,7 @@ class InteractiveCurveViewController
   struct CurveCursor {
     int curveIndex;
     int subCurveIndex;
-    float t;
+    double t;
   };
   // Closest vertical curve helper
   CurveCursor closestCurveIndexVertically(OMG::VerticalDirection direction,
